refactor: split "find aa changes" code into modules
In preparation for https://github.com/nextstrain/nextclade/pull/1454

Among other things, in order to render private and relative aa mutations we need to group them similarly to how absolute aa mutations are grouped. This involves finding adjacent mutations and nuc context for these mutations.

Sadly, the code for grouping is quite complex and is not immediately reusable (it combines aa mutation search and grouping at the same time), so let's do some refactoring.

Let's start from factoring away structs and functions that will be unchanged, to clear up some space for the action, to minimize diffs  and to reduce scrolling.