fix(web): error when selecting a cds fragment in annotation viz
The annotation visualization component were setting name of the CDS fragment as current CDS on click. This is incorrect: CDS name needs to be set instead (like it's done in the dropdown selector).

This lead to an error displayed in the peptide view when clicking on annotation viz and when name of CDS and CDS segment is different, which is always the case for Auspice datasets - Nextclade generates fragment names, as well as for multi-fragment CDSes.

Here I fix the problem by setting CDS name as current CDS on click, rather than fragment name.