fix: nuc mut context for relative aa muts
So far nuc context in aa mut groups used nuc muts relative to the true reference (precomputed after alignment). But in cases of aa mutations relative to other nodes, we need to either:
1. modify these nuc mutations
2. or to recompute them again, this time against another nodes' sequence
3. or to use private nuc mutations computed for this node.

The step 2 actually happens when we calculate relative nuc mutations so w could use those.

Bit I decided this time to opt for a hybrid of 2 & 3 and to 'recompute' mutations from alignment views. I am querying the node's aliment views only in the regions we are interested in (only in the groups' range). These views already contain the overlay of relative nuc mutations, so no actual recomputing is happening. This approach simplifies and clarifies code a little.