fix: ensure aa deletions participate in grouping
This required combining subs and dels and then `AaSub` and `AaDel` to be iterated as one iterator, sorted by position. I opted for iterator of `AbstractMutation`trait objects  - this required to make the trait object-safe, in this case moving a method returning `Self` into a separate trait.

It's a complicated way to do something simple, but I am happy that this solution avoids copying and leaves function interfaces mostly intact.