fix: prevent duplicate node names in output auspice json
When new nodes are placed onto a ref node, sometimes we create an intermediate new node as a copy of the existing ref node. They are all named the same - `{node_key}_internal`. This is insufficient to provide uniqueness when multiple new nodes are placed onto the same ref node.

To address that, here I change names to also include query name and index.