chore(infra): remove feature-policy header, remove interest-cohort from permissions-policy
Having both `feature-policy` and `permissions-policy` causes a "yellow error"  error/warning (warnor? errning?) because it's [deprecated](https://scotthelme.co.uk/goodbye-feature-policy-and-hello-permissions-policy/) (like if noone is using old browsers which don't know about it)

Same for `interest-cohort` entry in `permissions-policy`which requires a feature flag enabled in chrome to be functional:

> Error with Permissions-Policy header: Origin trial controlled feature not enabled: 'interest-cohort'.

Let's remove the outdated `feature-policy` header and remove `interest-cohort` entry from `permissions-policy`.

Still an "A" score on both
- https://securityheaders.com/?q=clades.nextstrain.org&followRedirects=on
- https://observatory.mozilla.org/analyze/clades.nextstrain.org
even though we allow `unsafe-eval` in order to be able to run wasm.

Security headers being a bit of a confusing mess nowadays (to put it lightly), which is always good for additional feeling of security of course! I am not even sure why I am doing this anymore. To get an "A" score I guess.