feat: add built-in ancestral search for clade founders
This adds a search for ancestor node with the same clade as the query sample. Let's call that "clade founder" node. Similarly this adds the same search but for custom clade-like attributes.

After the code is found we compute nuc and aa mutations relative to it.

This allows to avoid explicit repetitive ref node configuration in the ref tree.

The idea is similar to nearest mode/private mutations, and can be considered continuation of it: we start searching from the nearest node and ascend the tree towards the root, until we find a node which is closest to the root and having the same clade or clade-like attribute.