feat: switch auspice tree on dataset switch
Auspice state is global and is stored in Redux. From Nextclade we don't have access to how state is stored and operated on.

But we need to switch what Auspice is rendeing when viewed dataset is changed.

Instead of recomputing the state on every switch, I decided to try and precompute what's the necessary state in the webworker, as before but for each of the multiple datasets now. Then I dump the entire auspice state in recoil atom family (by dataset name). When a dataset is selected on tree page, I can now quickly restore the redux state from saved state.

This way we can hopefully quickly switch what Auspice is rendering.