fix: ensure trailing newline is included when parsing GFF3 region
Resolves https://github.com/nextstrain/nextclade/issues/1572

When parsing GFF3, we split up the file on `#sequence-region` pragma blocks. As per specification, there could be multiple of them, and we want to know about this.

The `bio` crate's `GffReader` does not support regions, so we do the splitting manually and then pass each region's slice to `bio` `GffReader`. However, due to mistakenly trimmed trailing newlines in the regions on our side, `bio` GFF3 parser would fail when it encounters a commented line as a last line of the region. The error and repro is described in https://github.com/nextstrain/nextclade/issues/1572.

In this PR I:
 - [x] make sure that the region content string is properly extracted, including the trailing newline character
 - [x] remove `.trim()` call such that this newline isn't removed

This way `bio` can understand our GFF3 blocks even when they have comments.