fix: ensure CDS segment strand is written
GFF: CDS strand is fixed in this commit. But  genes, to my understanding, don't have strand orientation, because they can consist of multiple CDS segments with different orientation. So genes I left genes with `.` instead of a particular strand value.

TBL: for CDS the range boundaries are correctly swapped, so no changes are needed. For genes, similarly to GFF, the strandedness is undefined, so I leave genes with forward orientation always.