refactor: use enum for truncation, rather than string matching
This is more type safe and less error-prone, because there is no need to synchronize exact string values in multiple places (which can be accidentally changed independently).

This also allows to retrieve truncation amount as an integer, should it be needed in the future (rather than string).

NOTE: the `truncation` field is currently set to the default  value`None` on all CDS segments in all places where reference annotation is used. And it sin only modified when query annotation is constructed. This may or may not be correct - ref annotation might need truncation to be calculated additionally. Though it is never used currently and typically reference annotation is supposed to be complete.