chore(ci): add gff validation of outptu query gff files to ci
Let's validate gff files produced during smoke tests. Smoke tests run on all datasets, so we cover a decent variety of outputs and this can help to catch some bugs early.

Currently the warnings due to missing `##sequence-region` pragmas are silenced. We are working on introducing them into the output.