fix(web): reset dataset suggestions fully when new data is added
Avoid leftover data which allows to run analysis in multi-dataset mode without re-running suggestions first.

Without rerunning suggestions on updated inputs we don't have a reliable mapping of sequence index to dataset name anymore, which can cause incorrect analysis results. I decided to not guess and forbid the analysis unless the suggestions are fresh. If inputs are modified, and auto-suggestions are disabled, user will see the "Suggest" button on multi-dataset tab, so that they are forced to re-run suggestions before they could the analysis.

We might be able to salvage the old suggestions if we really want to and the added data might simply go to "Unknown dataset", but it might take some effort (for starters, we need to detect when data is added, as opposed to replaced, fully or partially, or altered otherwise), and does not seem is not useful enough practically to spend efforts.