fix(web): reset file download status checkmarks on dataset switch
We could theoretically try and preserve the status until new analysis results, but probably more problems than its worth (need to know precise conditions when to reset).

Here I simply add the `key` prop to the component, so that there are multiple export components per dataset and they all have separate local state, which is reset on unmount.

This is consistent with behavior when switching pages. These  status icons are not meant to be long-lived anyways, just to give user an immediate feedback. In some UIs they also reset to normal state after a few seconds. But that requires some additional coding.