feat(web): add download svg button on tree page
This uses existing "SVG screenshot" functionality from Auspice's "Download data" dialog box.

The entire dialog box content is not necessarily relevant and working (it seems to be assuming metadata, charon server etc.), but this particular SVG functionality is useful and I thought I just add it as a separate button in the top right corner of the page.