fix(web): display suggestion error in an alert widget, rather than modal
Realization came after 0df1dc0 that we don't necessarily want global error modal on every suggestion engine failure, if at all. The reason is that suggestion functionality is optional and probably should not be blocking even if fails. So if a user drops an invalid fasta (file, link or paste) or sequences have no suggestions, this should not be a global error.

Here I try to introduce a separate error state for suggestion algo, and to display these errors, when they occur, unobtrusively, in an alert widget right under the suggestion card on main page, as well as on the datasets page.

Might have still missed some edge cases, but ux is a bit better now.