fix(web): crash when usgin single dataset from url
Resolves: #1631

Nextclade Web crashed due to a race between single-dataset initialization from url params
and periodic background re-fetch of datasets (background dataset updates to catch the new releases). The updates coming from the default dataset server would overwrite the custom dataset in the app state.

The first periodic fetch would happen right after the actual custom dataset was fetched, as main page starts to render, but after app component is done initialization. So manual runs would crash, but automatic runs with `?input-fasta` would not, because the main page is never rendered and skipped to the results page.

Additional nuissance is that adding sequences manually would trigger dataset suggestion algo and it would fail, because the custom dataset is obviously not present in the minimizer index and there are no matches possible. This manifested in an error alert to the users, even though running suggestions does not make sense in the first place.

Here I try to mitigate this situation, by setting  current dataset server url state atom to `undefined` whenever a custom dataset is requested with `?dataset-url=` or `?dataset-json-url=`. This way is nowhere to fetch the dataset index or data or minimizer index from, so:
 - periodic background dataset fetch knows to disable itself
 - dataset suggestion algo is skipped

This has a defect - once launched with a custom dataset, the app enters a state where there is only this dataset and nothing else. But I think it was the same in previous releases too. The workaround is to refresh the browser page (or open the app without url params). I'd like to address this in a separate PR: <insert issue number here>

Overall, the initialization code has become a bit of a spaghetti at this point, and we need to refactor and disentangle the init and app state at some point in order prevent these kind of edge cases from happening.