fix(web): ensure input promises are resolved before passing to worker
The helper function `promiseAllObject()` had an unsafe cast which has hidden the unresolved `Promise`s on the output object.

This would cause a crash when object is later passed to a WebWorker (`Promises` are not serializable using the structured clone algorithm).

Let's use a safer helper from `p-promise` package. It's from Sindre, so must e good.

I will try to remove the `promiseAllObject()` in all other places in the followup commits.