fix: calculate seed cover correctly when query is shorter than reference
the seed matching algorithm requires that a certain fraction of the query sequences
is covered by seed matches. When the reference is shorter than the query sequence,
the fraction that can be covered by seed matches is limited by the length of the
reference sequence. Here we change the calculation of the fraction covered by seeds
to the fraction of the sum of all seed length and minimum of qry and ref length.
This won't change the behavior for datasets where the reference is complete, but
results in the intended behavior when partial sequences are used as references.
Previously, dataset maintainers had to set the min_seed_cover to unreasonably low
values in cases the reference was much short than the queries.