fix: ensure canonical order of CSV columns
Columns in CSV and TSV files will always be rendered in their predefined canonical order, even when using `--output-columns-selection` or column config in web.

Currently it is:
- if `--output-columns-selection` is not used: canonical order
- if `--output-columns-selection` is used: columns from categories in their requested order , then individual columns in their requested order

I think this is not very intuitive, and is particularly inconcenient in cases like:
```
--output-columns-selection=index,seqName,priv-muts
```
where private mutations (category based) are listed before sequence name (individual column). And although it would be nice to sort everything in requested order across columns and categories, it would currently require some more effort and bookkeeping, so I oped for a simpler solution of always enforcing canonical order.